urgh, cast to int

+++ src/main/java/fr/tunaki/stackoverflow/burnaki/bot/command/StopCommand.java
@@ -56,7 +56,7 @@ public class StopCommand implements Command {
 			room.send("Nope. Only a moderator, a room owner or Tunaki can stop me.");
 			return;
 		}
-		long roomId = room.getRoomId();
+		int roomId = (int) room.getRoomId();
 		if (roomId == properties.getHqRoomId()) {
 			burnaki.getBurnRooms().forEach((id, r) -> r.getRoom().send("I'm stopping, see you guys later!"));
 			room.send("Bye.").thenRun(context::close);
